requirements.txt
fastapi[standard]~=0.115.12
python-dotenv~=1.1.0
firebase-admin~=6.8.0

pydantic~=2.11.5
starlette~=0.46.2